+++ examples/kodo_helpers.py
@@ -212,41 +212,47 @@ class EncodeStateViewer(StateViewer):
 
     """Class for displaying the encoding coefficients."""
 
-    def __init__(self, size, canvas, canvas_position=(0, 0)):
+    def __init__(self, size, canvas, canvas_position=(0, 0),
+                 wrap_around=False):
         """Create EncodeStateViewer."""
         super(EncodeStateViewer, self).__init__(
             size, canvas, canvas_position)
+        # if true, wrap_around, else push up.
+        self.wrap_around = wrap_around
 
     def trace_callback(self, zone, message):
         """Callback to be used with the encoder trace API."""
-
         if zone == "set_symbols":
-            self.symbols = int(message.split('\n')[-3].split()[0]) + 1
+            for line in reversed(message.split('\n')):
+                elements = line.split()
+                if len(elements) > 1 and elements[1] == 'I:':
+                    self.symbols = int(elements[0]) + 1
+                    break
+
             self.state = [[] for i in range(self.symbols)]
             self.index = 0
             return
 
-        if zone == "symbol_index_before_write_uncoded_symbol":
+        if zone == "symbol_index_after_write_uncoded_symbol":
             index = int(message.split(' ')[-1])
             symbol = [0 for i in range(self.symbols)]
             symbol[index] = 1
 
-        elif zone == "coefficients_after_write_symbol":
+        elif zone == "symbol_coefficients_after_write_symbol":
             symbol = message[3:].split(' ')[:-1]
             symbol = map(int, symbol)
         else:
             return
 
-        # Push up
         if self.index < self.symbols:
             self.state[self.index] = symbol
         else:
             self.state = self.state[1:] + [symbol]
         self.index += 1
 
-        # Wrap around
-        #self.state[self.index] = symbol
-        #self.index = (self.index + 1) % self.symbols
+        if self.wrap_around:
+            self.index = self.index % self.symbols
+            self.state[self.index-1] = symbol
 
         self.show_decode_state(self.state)